+++ src/ivfflat.h
@@ -315,7 +315,7 @@ VectorArraySet(VectorArray arr, int offset, Pointer val)
 /* Methods */
 VectorArray VectorArrayInit(int maxlen, int dimensions, Size itemsize);
 void		VectorArrayFree(VectorArray arr);
-Size		IvfflatKmeansMemoryRequired(int numSamples, int numCenters, int dimensions, Size itemsize, int samplesCapacity);
+Size		IvfflatKmeansMemoryRequired(int numSamples, int numCenters, int dimensions, Size itemsize, Size samplesCapacity);
 int			IvfflatMaxKmeansSamples(int numCenters, int dimensions, Size itemsize);
 void		IvfflatKmeans(Relation index, VectorArray samples, VectorArray centers, const IvfflatTypeInfo * typeInfo);
 FmgrInfo   *IvfflatOptionalProcInfo(Relation index, uint16 procnum);